+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enhancements
     - Default configuration entries for files and urls are available through `RefInspector.Config.default_files/0` and `RefInspector.Config.default_urls/0`
     - Output of mix task `ref_inspector.download` can be prevented by passing `--quiet` upon invocation. This does NOT imply `--force` and will still ask for confirmation
+    - The library used to download the database files can be changed by configuring a module implementing the `RefInspector.Downloader.Adapter` behaviour
     - The library used to read YAML files can be changed by using the `:yaml_file_reader` configuration
 
 ## v1.0.0 (2018-11-24)